cutoptix-backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.0.0
django-filter==23.1
psycopg2-binary==2.9.6
celery==5.2.7
redis==4.5.4
pillow==9.5.0
reportlab==4.0.4
openpyxl==3.1.2
python-decouple==3.8
gunicorn==20.1.0
whitenoise==6.4.0
pytest-django==4.5.2
pytest==7.3.1
factory-boy==3.2.1
django-extensions==4.1
